- Grouped limiters in one .h and one .cpp - Grouped damping_methods in one .h and one .cpp - Removed separate factories. Made SolverFactory generic by introducing template parameters. - Made usage of boost::shared_ptr instead of own pointer handling. - Adapted CMakeLists.txt according to changes.
- Split parameter enforce_limits into enforce_pos_limits and enforce_vel_limits
