+++ cob_twist_controller/cfg/TwistController.cfg
@@ -17,23 +17,23 @@ constraints_enum = gen.enum([
                        gen.const("WLN_JLA",        int_t, 2, "Weighted-least-norm joint-limit-avoidance")],
                      "enum types for the constraints")
 
-gen.add("damping_method",    int_t,    0, "The damping method to use (edited via an enum)", 0, None, None, edit_method=damping_method_enum)
+gen.add("damping_method",     int_t,    0, "The damping method to use (edited via an enum)", 0, None, None, edit_method=damping_method_enum)
+gen.add("damping_factor",     double_t, 0, "The constant damping_factor (used in CONSTANT)",  0.01, 0, 1)
+gen.add("lambda0",            double_t, 0, "Value for maximum damping_factor (used in MANIPULABILITY)",  0.1, 0, 10)
+gen.add("wt",                 double_t, 0, "Value for manipulability threshold (used in MANIPULABILITY)",  0.005, 0, 1)
 
-gen.add("eps",               double_t, 0, "Value for singular value threshold (used for truncation)",    0.001, 0,   1)
+gen.add("constraint",         int_t,    0, "The constraint to use (edited via an enum)", 2, None, None, edit_method=constraints_enum)
 
-gen.add("damping_factor",    double_t, 0, "The constant damping_factor (used in CONSTANT)",  0.01, 0, 1)
-gen.add("lambda0",           double_t, 0, "Value for maximum damping_factor (used in MANIPULABILITY)",  0.1, 0, 10)
-gen.add("wt",                double_t, 0, "Value for manipulability threshold (used in MANIPULABILITY)",  0.005, 0, 1)
+gen.add("eps",                double_t, 0, "Value for singular value threshold (used for truncation)",    0.001, 0,   1)
 
-gen.add("constraint",        int_t,    0, "The constraint to use (edited via an enum)", 2, None, None, edit_method=constraints_enum)
+gen.add("keep_direction",     bool_t,   0, "With keep_direction the whole joint positions and velocities vector is affected by a scaling factor. Else only individual components of the vectors are affected -> direction will be changed.", True)
+gen.add("enforce_vel_limits", bool_t,   0, "If 'True', enforce joint velocity limits in TwistController, else let RobotHW handle the limits",  True)
+gen.add("enforce_pos_limits", bool_t,   0, "If 'True', enforce joint position limits in TwistController, else let RobotHW handle the limits",  True)
+gen.add("tolerance",          double_t, 0, "Tolerance around joint position limits where velocities are normalized [degrees]",  5, 0, 15)
 
-gen.add("keep_direction",    bool_t,   0, "With keep_direction the whole joint positions and velocities vector is affected by a scaling factor. Else only individual components of the vectors are affected -> direction will be changed.", True)
-gen.add("enforce_limits",    bool_t,   0, "If 'True', enforce joint position limits and normalize velocities in TwistController, else let RobotHW handle the limits",  True)
-gen.add("tolerance",         double_t, 0, "Tolerance around joint position limits where velocities are normalized [degrees]",  5, 0, 15)
-
-gen.add("base_compensation", bool_t,   0, "Compensate base motion",  False)
-gen.add("base_active",       bool_t,   0, "Command base actively",  False)
-gen.add("base_ratio",        double_t, 0, "Value for base_ratio",  0.0, 0, 1)
-gen.add("reset_markers",     bool_t,   0, "Reset Markers",  False)
+gen.add("base_compensation",  bool_t,   0, "Compensate base motion",  False)
+gen.add("base_active",        bool_t,   0, "Command base actively",  False)
+gen.add("base_ratio",         double_t, 0, "Value for base_ratio",  0.0, 0, 1)
+gen.add("reset_markers",      bool_t,   0, "Reset Markers",  False)
 
 exit(gen.generate(PACKAGE, "cob_twist_controller", "TwistController"))
