+++ cob_twist_controller/include/cob_twist_controller/constraint_solvers/solvers/constraint_solver_base.h
@@ -62,10 +62,10 @@ class ConstraintSolver
 
     protected:
 
-        ConstraintSolver(AugmentedSolverParams &asSolverParams,
+        ConstraintSolver(AugmentedSolverParams &asParams,
                          Matrix6Xd &jacobianData,
                          Eigen::Transpose<Matrix6Xd> &jacobianDataTransposed)
-                         : asSolverParams_(asSolverParams),
+                         : asParams_(asParams),
                            jacobianData_(jacobianData),
                            jacobianDataTransposed_(jacobianDataTransposed),
                            dampingFactor_(0.0)
@@ -81,7 +81,7 @@ class ConstraintSolver
         Eigen::MatrixXd calculatePinvJacobianBySVD(Eigen::JacobiSVD<Eigen::MatrixXd> svd) const;
 
 
-        const AugmentedSolverParams &asSolverParams_; ///< References the augmented solver parameters.
+        const AugmentedSolverParams &asParams_; ///< References the augmented solver parameters.
         const Matrix6Xd &jacobianData_; ///< References the current Jacobian (matrix data only).
         const Eigen::Transpose<Matrix6Xd> &jacobianDataTransposed_; ///< References the current Jacobian transpose (matrix data only).
         double dampingFactor_; ///< The currently set damping factor.
