+++ cob_twist_controller/include/cob_twist_controller/constraint_solvers/solvers/weighted_least_norm_solver.h
@@ -44,10 +44,10 @@ class WeightedLeastNormSolver : public ConstraintSolver
          */
         virtual Eigen::MatrixXd solve(const Eigen::VectorXd &inCartVelocities, const KDL::JntArray& q, const KDL::JntArray& last_q_dot) const;
 
-        WeightedLeastNormSolver(AugmentedSolverParams &asSolverParams,
+        WeightedLeastNormSolver(AugmentedSolverParams &asParams,
                                   Matrix6Xd &jacobianData,
                                   Eigen::Transpose<Matrix6Xd> &jacobianDataTransposed)
-                                  : ConstraintSolver(asSolverParams,
+                                  : ConstraintSolver(asParams,
                                                      jacobianData,
                                                      jacobianDataTransposed)
         {
